requirements.txt
## Setup py requirements
jsonpickle==0.9.6
six==1.11.0
redis==2.10.6
fakeredis==0.10.3
future==0.16.0
numpy==1.14.5
typing==3.6.4
ruamel.yaml==0.15.37
requests~=2.20
keras==2.1.6
tensorflow==1.10.0
h5py==2.7.1
apscheduler==3.5.1
tqdm==4.23.3
ConfigArgParse==0.13.0
networkx==2.1
pykwalify<=1.6.0
coloredlogs==10.0
flask==1.0.2
flask_cors==3.0.4
scikit-learn==0.19.1
rasa_nlu~=0.13.0
colorhash==1.0.2
pika==0.11.2
jsonschema==2.6.0
packaging==17.1
gevent==1.2.2
pyyaml==3.12
pytz==2018.4
rasa-core-sdk~=0.12.1
rasa_core~=0.12.0
pymongo==3.5.1
python-dateutil==2.7.3
spacy==2.0.12
sklearn_crfsuite==0.3.6
msgpack==0.5.6
weather-api==1.0.7